Fix animation on transaction status screen
